refactor: 拆分行情展示与运行时模块
