+++ scripts/run_checks.py
@@ -12,6 +12,7 @@ COMPILE_TARGETS = (
     "goldmonitor/daily_digest.py",
     "goldmonitor/market_adapters.py",
     "goldmonitor/market_data.py",
+    "goldmonitor/market_runtime.py",
     "goldmonitor/data_archive.py",
     "goldmonitor/portfolio_analytics.py",
     "goldmonitor/review_notes.py",
@@ -38,6 +39,7 @@ COMPILE_TARGETS = (
     "tests/test_risk_analysis_module.py",
     "tests/test_market_data_module.py",
     "tests/test_market_adapters_module.py",
+    "tests/test_market_runtime_module.py",
     "tests/test_market_adapters_app.py",
     "tests/test_market_source_management.py",
     "tests/test_data_archive_module.py",
@@ -87,6 +89,7 @@ PYTHON_CHECKS = (
     "tests/test_risk_analysis_module.py",
     "tests/test_market_data_module.py",
     "tests/test_market_adapters_module.py",
+    "tests/test_market_runtime_module.py",
     "tests/test_settings_store_module.py",
     "tests/test_notifications_module.py",
     "tests/test_scheduler_module.py",
