+++ templates/index.html
@@ -672,6 +672,7 @@
 <script src="https://cdn.jsdelivr.net/npm/chart.js@4.4.0/dist/chart.umd.min.js"></script>
 <script src="https://cdn.socket.io/4.7.2/socket.io.min.js"></script>
 <script src="/static/app-shell.js?v={{ app_version }}"></script>
+<script src="/static/market-dashboard.js?v={{ app_version }}"></script>
 
 <script src="/static/settings-center.js?v={{ app_version }}"></script>
 <script src="/static/operations-center.js?v={{ app_version }}"></script>
