+++ tests/contract_checks.ps1
@@ -192,7 +192,7 @@ Assert-Contains -Path "goldmonitor\market_adapters.py" -Pattern 'class MarketAda
 Assert-Contains -Path "goldmonitor\market_adapters.py" -Pattern 'def fetch_http_source' -Message "market adapters must expose the shared HTTP fetch helper"
 Assert-Contains -Path "app.py" -Pattern 'def build_market_adapter_registry' -Message "app.py must build the configured market adapter registry"
 Assert-Contains -Path "app.py" -Pattern 'def get_market_adapter_catalog' -Message "app.py must expose the market adapter catalog"
-Assert-Contains -Path "app.py" -Pattern 'state\["adapters"\]\s*=\s*adapters' -Message "source health must include the market adapter catalog"
+Assert-Contains -Path "goldmonitor\market_runtime.py" -Pattern 'state\["adapters"\]\s*=\s*adapters' -Message "source health must include the market adapter catalog"
 Assert-Contains -Path "app.py" -Pattern 'SOURCE_METRICS_PATH' -Message "source rolling metrics must persist locally"
 Assert-Contains -Path "goldmonitor\market_data.py" -Pattern 'class SourceMetricsStore' -Message "market data must expose the rolling metrics store"
 Assert-Contains -Path "goldmonitor\market_data.py" -Pattern 'success_rate_pct' -Message "source health must expose rolling success rate"
@@ -231,6 +231,8 @@ Assert-Contains -Path "app.py" -Pattern 'def fetch_sina_gold_result' -Message "a
 Assert-Contains -Path "app.py" -Pattern 'def fetch_eastmoney_gold_result' -Message "app.py must preserve the compatible Eastmoney fetch entry"
 Assert-Contains -Path "app.py" -Pattern 'def fetch_goldprice_data_result' -Message "app.py must preserve the compatible GoldPrice fetch entry"
 Assert-Contains -Path "app.py" -Pattern 'def fetch_market_data_result' -Message "app.py must preserve the compatible market fetch entry"
+Assert-Contains -Path "goldmonitor\market_runtime.py" -Pattern 'class MarketRuntime' -Message "market runtime module must own price refresh orchestration"
+Assert-Contains -Path "goldmonitor\market_runtime.py" -Pattern 'def fetch_market_data_result' -Message "market runtime module must own market source orchestration"
 Assert-Contains -Path "app.py" -Pattern 'parse_sina_forex' -Message "app.py must parse Sina USD/CNY rates"
 Assert-Contains -Path "app.py" -Pattern 'parse_frankfurter_forex' -Message "app.py must parse Frankfurter USD/CNY rates"
 Assert-Contains -Path "app.py" -Pattern 'rate_cached' -Message "price updates must expose cached rate status"
@@ -347,7 +349,7 @@ Assert-Contains -Path "app.py" -Pattern 'PRICE_HISTORY_PATH' -Message "app.py mu
 Assert-Contains -Path "goldmonitor\socket_history_review.py" -Pattern '@socketio\.on\("get_price_history"\)' -Message "backend must expose price history lookup"
 Assert-Contains -Path "goldmonitor\socket_history_review.py" -Pattern '@socketio\.on\("export_price_history"\)' -Message "backend must expose price history export"
 Assert-Contains -Path "templates\index.html" -Pattern 'openHistory' -Message "frontend must expose history review"
-Assert-Contains -Path "static\app.js" -Pattern 'requestChartHistory' -Message "frontend chart periods must request historical windows"
+Assert-Contains -Path "static\market-dashboard.js" -Pattern 'requestChartHistory' -Message "frontend chart periods must request historical windows"
 Assert-Contains -Path "templates\index.html" -Pattern 'data-period="4h"' -Message "frontend chart must expose a 4-hour period"
 Assert-Contains -Path "templates\index.html" -Pattern 'data-period="7d"' -Message "frontend chart must expose a 7-day period"
 Assert-Contains -Path "templates\index.html" -Pattern 'data-period="30d"' -Message "frontend chart must expose a 30-day period"
@@ -416,12 +418,12 @@ Assert-Contains -Path "templates\index.html" -Pattern 'settingsTabOps' -Message 
 Assert-Contains -Path "templates\index.html" -Pattern 'exportDiagnostics' -Message "frontend must export diagnostics"
 Assert-Contains -Path "templates\index.html" -Pattern 'openExportsFolder' -Message "frontend must expose export folder opening"
 Assert-Contains -Path "templates\index.html" -Pattern 'id="priceRetry"' -Message "index.html must show a manual price retry button"
-Assert-Contains -Path "static\app.js" -Pattern 'socket\.emit\(''refresh_price''' -Message "index.html must request manual price refresh"
-Assert-Contains -Path "static\app.js" -Pattern 'socket\.on\(''fetch_status''' -Message "index.html must render fetch status"
-Assert-Contains -Path "static\app.js" -Pattern 'socket\.connect\(\)' -Message "index.html must provide a manual reconnect action"
+Assert-Contains -Path "static\market-dashboard.js" -Pattern 'socket\.emit\(''refresh_price''' -Message "index.html must request manual price refresh"
+Assert-Contains -Path "static\market-dashboard.js" -Pattern 'socketClient\.on\(''fetch_status''' -Message "index.html must render fetch status"
+Assert-Contains -Path "static\market-dashboard.js" -Pattern 'socket\.connect\(\)' -Message "index.html must provide a manual reconnect action"
 Assert-Contains -Path "static\app.js" -Pattern 'socket\.on\(''connect_error''' -Message "index.html must explain initial connection failures"
-Assert-Contains -Path "static\app.js" -Pattern 'rate_cached' -Message "index.html must label cached forex rates"
-Assert-Contains -Path "static\app.js" -Pattern 'gold_cached' -Message "index.html must preserve cached gold status"
+Assert-Contains -Path "static\market-dashboard.js" -Pattern 'rate_cached' -Message "index.html must label cached forex rates"
+Assert-Contains -Path "static\market-dashboard.js" -Pattern 'gold_cached' -Message "index.html must preserve cached gold status"
 Assert-Contains -Path "GoldMonitor.spec" -Pattern 'exclude_binaries=True' -Message "PyInstaller must build in onedir mode"
 Assert-Contains -Path "GoldMonitor.spec" -Pattern 'COLLECT\(' -Message "PyInstaller spec must collect files into a distribution directory"
 Assert-Contains -Path "installer\GoldMonitor.iss" -Pattern 'PrivilegesRequired=lowest' -Message "installer must be per-user"
