+++ tests/test_frontend_market_status_contract.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 
 def test_frontend_fetch_status_uses_degraded_status_contract():
-    js = Path("static/app.js").read_text(encoding="utf-8")
+    js = Path("static/market-dashboard.js").read_text(encoding="utf-8")
     start = js.index("function applyFetchStatus")
     end = js.index("function refreshPrice", start)
     body = js[start:end]
