+++ tests/test_run_checks_script.py
@@ -9,6 +9,7 @@ EXPECTED_COMPILE_TARGETS = (
     "goldmonitor/daily_digest.py",
     "goldmonitor/market_adapters.py",
     "goldmonitor/market_data.py",
+    "goldmonitor/market_runtime.py",
     "goldmonitor/data_archive.py",
     "goldmonitor/portfolio_analytics.py",
     "goldmonitor/review_notes.py",
@@ -35,6 +36,7 @@ EXPECTED_COMPILE_TARGETS = (
     "tests/test_risk_analysis_module.py",
     "tests/test_market_data_module.py",
     "tests/test_market_adapters_module.py",
+    "tests/test_market_runtime_module.py",
     "tests/test_market_adapters_app.py",
     "tests/test_market_source_management.py",
     "tests/test_data_archive_module.py",
@@ -84,6 +86,7 @@ EXPECTED_PYTHON_CHECKS = (
     "tests/test_risk_analysis_module.py",
     "tests/test_market_data_module.py",
     "tests/test_market_adapters_module.py",
+    "tests/test_market_runtime_module.py",
     "tests/test_settings_store_module.py",
     "tests/test_notifications_module.py",
     "tests/test_scheduler_module.py",
